# contracts/openapi/control-plane.v0.yaml
openapi: 3.1.0
info:
  title: Waylorn control-plane API (DRAFT)
  version: 0.1.0-draft
  description: |
    DRAFT contract proposed by the frontend (ADR 0008). The .NET control plane does not
    implement it yet. When the backend exists it emits the authoritative document and this
    file is replaced.

    Invariants encoded here:
    - Every path is tenant-scoped by organization; the backend verifies the caller's scope.
    - `permittedActions` fields are usability hints computed by the backend. The backend
      re-evaluates RBAC + ABAC on every request; clients must not treat hints as authority.
    - Operational values carry provenance (`observedAt` / `computedAt`, `source`, method or
      model version). Analytics and ML values are never computed by clients.
    - Errors use RFC 9457 problem details with a `correlationId`.
servers:
  - url: /api/v0
security:
  - bearer: []
...
paths:
  /me:
    get:
      tags: [identity]
      operationId: getMe
      summary: Current principal and organizations the principal may access.
      responses:
        '200':
          description: Principal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Principal' }
        default: { $ref: '#/components/responses/Problem' }
...
  /orgs/{orgId}/sites:
    get:
      tags: [tenancy]
      operationId: listSites
      parameters: [{ $ref: '#/components/parameters/OrgId' }]
      responses:
        '200':
          description: Sites visible to the caller
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Site' }
        default: { $ref: '#/components/responses/Problem' }
  /orgs/{orgId}/hierarchy:
    get:
      tags: [tenancy]
      operationId: listHierarchyChildren
      summary: Children of a hierarchy node (organization → region → site → zone → line).
      parameters:
        - { $ref: '#/components/parameters/OrgId' }
        - name: parentId
          in: query
          description: Omit for top-level regions.
          schema: { type: string }
      responses:
        '200':
          description: Child nodes
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/HierarchyNode' }
        default: { $ref: '#/components/responses/Problem' }
  /orgs/{orgId}/assets:
    get:
      tags: [assets]
      operationId: listAssets
      parameters:
        - { $ref: '#/components/parameters/OrgId' }
        - { name: siteId, in: query, schema: { type: string } }
        - { name: zoneId, in: query, schema: { type: string } }
        - { name: lineId, in: query, schema: { type: string } }
        - name: kind
          in: query
          schema: { $ref: '#/components/schemas/AssetKind' }
        - name: health
          in: query
          schema: { $ref: '#/components/schemas/HealthState' }
        - name: lifecycle
          in: query
          schema: { $ref: '#/components/schemas/LifecycleState' }
        - name: q
          in: query
          description: Free-text search over name, tag, serial, manufacturer and model.
          schema: { type: string, maxLength: 200 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, tag, health, lastSeen, installedYear]
            default: tag
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: One page of assets
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetPage' }
        default: { $ref: '#/components/responses/Problem' }
  /orgs/{orgId}/assets/{assetId}:
    get:
      tags: [assets]
      operationId: getAsset
      parameters:
        - { $ref: '#/components/parameters/OrgId' }
        - { $ref: '#/components/parameters/AssetId' }
      responses:
        '200':
          description: Asset detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetDetail' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema: { type: string }
    AssetId:
      name: assetId
      in: path
      required: true
      schema: { type: string }
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from `page.nextCursor`.
      schema: { type: string, maxLength: 512 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status, correlationId]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        correlationId: { type: string }
        code:
          type: string
          description: Stable machine-readable reason, e.g. `step_up_required`, `forbidden_scope`.
        stepUp: { $ref: '#/components/schemas/StepUpRequirement' }
    PageInfo:
      type: object
      properties:
        nextCursor: { type: string }
        totalEstimate:
          type: integer
          description: Approximate count; may be omitted for expensive queries.
    Provenance:
      type: object
      required: [source]
      properties:
        source:
          type: string
          description: System or adapter that produced the value, e.g. `site-agent/opcua`, `analytics/r-reliability`.
        observedAt: { type: string, format: date-time }
        computedAt: { type: string, format: date-time }
    SafetyClass:
      type: string
      enum: [GREEN, AMBER, RED]
    Environment:
      type: string
      enum: [production, staging, lab, development]
    HealthState:
      type: string
      enum: [ok, warning, fault, unknown]
...
    ConnectivityState:
      type: string
      enum: [connected, degraded, disconnected, unknown]
    LifecycleState:
      type: string
      enum: [unknown, planned, commissioning, in_service, maintenance, standby, decommissioned]
    AssetKind:
      type: string
      enum: [IndustrialAsset, ComputeAsset, NetworkAsset, CloudResource, StorageAsset, ApplicationAsset, SecurityAsset]
...
    ComputePlatform:
      type: string
      enum: [bare_metal, virtual_machine, kubernetes_node, podman_host, cloud_instance]
    InfraProvider:
      type: string
      enum: [on_premises, hetzner, aws, azure, gcp, private_cloud, customer_datacenter]
    IdentityProviderRef:
      type: object
      required: [id, displayName, protocol]
      properties:
        id: { type: string }
        displayName: { type: string }
        protocol: { type: string, enum: [oidc, saml, ldap, local] }
    OrganizationRef:
      type: object
      required: [id, slug, name]
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
    Principal:
      type: object
      required: [subject, displayName, identityProvider, organizations]
      properties:
        subject: { type: string }
        displayName: { type: string }
        email: { type: string }
        identityProvider: { $ref: '#/components/schemas/IdentityProviderRef' }
        organizations:
          type: array
          items:
            allOf:
              - { $ref: '#/components/schemas/OrganizationRef' }
              - type: object
                required: [roles]
                properties:
                  roles:
                    type: array
                    description: Role names for display only; authorization is evaluated server-side.
                    items: { type: string }
    SiteConnectivity:
      type: object
      required: [state]
      properties:
        state: { $ref: '#/components/schemas/ConnectivityState' }
        lastContactAt: { type: string, format: date-time }
    Site:
      type: object
      required: [id, code, name, regionName, environment, timezone, connectivity]
      properties:
        id: { type: string }
        code: { type: string, description: "Short plant code shown in context bars, e.g. `DE-MUC-P3`." }
        name: { type: string }
        regionName: { type: string }
        environment: { $ref: '#/components/schemas/Environment' }
        timezone: { type: string, description: "IANA timezone" }
        connectivity: { $ref: '#/components/schemas/SiteConnectivity' }
    HierarchyNode:
      type: object
      required: [id, level, name, childCount]
      properties:
        id: { type: string }
        level: { type: string, enum: [region, site, zone, line] }
        name: { type: string }
        code: { type: string }
        parentId: { type: string }
        childCount: { type: integer }
        assetCount: { type: integer }
    TenancyContext:
      type: object
      description: Full location of a resource. Rendered on every consequential dialog.
      required: [organization, site]
      properties:
        organization: { $ref: '#/components/schemas/OrganizationRef' }
        region: { type: string }
        site:
          type: object
          required: [id, code, name, environment]
          properties:
            id: { type: string }
            code: { type: string }
            name: { type: string }
            environment: { $ref: '#/components/schemas/Environment' }
        zone:
          type: object
          required: [id, name]
          properties:
            id: { type: string }
            name: { type: string }
        line:
          type: object
          required: [id, name]
          properties:
            id: { type: string }
            name: { type: string }
...
    HealthSummary:
      type: object
      required: [state]
      properties:
        state: { $ref: '#/components/schemas/HealthState' }
        reason: { type: string }
        observedAt: { type: string, format: date-time }
    AssetConnectivity:
      type: object
      required: [state]
      properties:
        state: { $ref: '#/components/schemas/ConnectivityState' }
        lastSeenAt: { type: string, format: date-time }
        path:
          type: string
          description: Human-readable collection path, e.g. `Serial RS-232 → Moxa NPort → site agent`.
    AssetSummary:
      type: object
      required: [id, kind, name, tag, context, lifecycle, health, connectivity, identityConfidence]
      properties:
        id: { type: string, description: "Stable opaque identifier." }
        kind: { $ref: '#/components/schemas/AssetKind' }
        name: { type: string }
        tag: { type: string, description: "Plant tag when known; otherwise a system-generated `UNASSIGNED-{id}` label." }
        manufacturer: { type: string }
        model: { type: string }
        installedYear: { type: integer }
        context: { $ref: '#/components/schemas/TenancyContext' }
        lifecycle: { $ref: '#/components/schemas/LifecycleState' }
        health: { $ref: '#/components/schemas/HealthSummary' }
        connectivity: { $ref: '#/components/schemas/AssetConnectivity' }
        identityConfidence:
          type: string
          enum: [confirmed, probable, unverified]
          description: Reconciliation state. Discovered identifiers are claims until confirmed.
    AssetPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AssetSummary' }
        page: { $ref: '#/components/schemas/PageInfo' }
        facets:
          type: object
          properties:
            kind:
              type: array
              items: { $ref: '#/components/schemas/FacetCount' }
            health:
              type: array
              items: { $ref: '#/components/schemas/FacetCount' }
    FacetCount:
      type: object
      required: [value, count]
      properties:
        value: { type: string }
        count: { type: integer }
    ObservedValue:
      type: object
      required: [value, provenance]
      properties:
        value: { type: string }
        provenance: { $ref: '#/components/schemas/Provenance' }
    Capability:
      type: object
      description: |
        Capability of an asset, with four independent states. Discovery (`observed`) never
        implies authorization.
      required: [operation, label, safetyClass, observed, declared, siteApproved, currentlyAuthorized]
      properties:
        operation: { type: string }
        label: { type: string }
        safetyClass: { $ref: '#/components/schemas/SafetyClass' }
        observed: { type: boolean }
        declared: { type: boolean }
        siteApproved: { type: boolean }
        currentlyAuthorized: { type: boolean }
    ProtocolBinding:
      type: object
      required: [protocol, role, source]
      properties:
        protocol:
          type: string
          description: e.g. `OPC UA`, `Modbus RTU`, `S7comm`, `EtherNet/IP`, `BACnet/IP`, `MQTT Sparkplug B`, `Proprietary serial`.
        version: { type: string }
        role: { type: string, enum: [server, client, publisher, subscriber, passive] }
        interfaceId: { type: string }
        source: { type: string, enum: [observed, declared] }
    AssetInterface:
      type: object
      required: [id, medium, label]
      properties:
        id: { type: string }
        medium: { type: string, enum: [ethernet, serial_rs232, serial_rs485, fieldbus, wireless, virtual] }
        label: { type: string }
        networkZone: { type: string }
        addressRef:
          type: string
          description: Redacted or tokenised address; raw addresses stay in the site store.
        gatewayPath:
          type: array
          items: { type: string }
    Owner:
      type: object
      required: [team]
      properties:
        team: { type: string }
        contact: { type: string }
    VendorAttribute:
      type: object
      description: Vendor-specific detail preserved verbatim alongside the common model.
      required: [namespace, key, label, value]
      properties:
        namespace: { type: string, description: "e.g. `siemens.s7`, `rockwell.logix`, `legacy.serial`." }
        key: { type: string }
        label: { type: string }
        value: { type: string }
        unit: { type: string }
    ExternalReference:
      type: object
      required: [system, externalId]
      properties:
        system: { type: string, description: "e.g. `SAP PM`, `Maximo`, `ServiceNow CMDB`." }
        externalId: { type: string }
        syncedAt: { type: string, format: date-time }
        href: { type: string }
    IndustrialExtension:
      type: object
      required: [kind, controlRole, vendorSupport]
      properties:
        kind: { type: string, const: IndustrialAsset }
        controlRole: { type: string, enum: [plc, hmi, drive, machine, sensor, gateway, robot, safety_controller, other] }
        vendorSupport: { type: string, enum: [supported, limited, end_of_life, unknown] }
        programmingEnvironment: { type: string }
        protocolsSummary: { type: string }
    ComputeExtension:
      type: object
      required: [kind, platform, provider]
      properties:
        kind: { type: string, const: ComputeAsset }
        platform: { $ref: '#/components/schemas/ComputePlatform' }
        provider: { $ref: '#/components/schemas/InfraProvider' }
        operatingSystem: { type: string }
        cpuCores: { type: integer }
        memoryBytes: { type: integer, format: int64 }
    NetworkExtension:
      type: object
      required: [kind, role]
      properties:
        kind: { type: string, const: NetworkAsset }
        role: { type: string, enum: [switch, router, firewall, data_diode, wireless_ap, serial_server, other] }
        portCount: { type: integer }
        managed: { type: boolean }
    CloudExtension:
      type: object
      required: [kind, provider, account, resourceType]
      properties:
        kind: { type: string, const: CloudResource }
        provider: { $ref: '#/components/schemas/InfraProvider' }
        account: { type: string }
        region: { type: string }
        resourceType: { type: string }
    StorageExtension:
      type: object
      required: [kind, storageType]
      properties:
        kind: { type: string, const: StorageAsset }
        storageType: { type: string, enum: [block, file, object, database] }
        capacityBytes: { type: integer, format: int64 }
        usedBytes: { type: integer, format: int64 }
        worm: { type: boolean }
    ApplicationExtension:
      type: object
      required: [kind, runtime]
      properties:
        kind: { type: string, const: ApplicationAsset }
        runtime: { type: string, enum: [kubernetes, podman, systemd, windows_service, saas, other] }
        version: { type: string }
    SecurityExtension:
      type: object
      required: [kind, function]
      properties:
        kind: { type: string, const: SecurityAsset }
        function: { type: string, enum: [firewall, ids, siem_collector, certificate_authority, jump_host, other] }
    AssetExtension:
      oneOf:
        - { $ref: '#/components/schemas/IndustrialExtension' }
        - { $ref: '#/components/schemas/ComputeExtension' }
        - { $ref: '#/components/schemas/NetworkExtension' }
        - { $ref: '#/components/schemas/CloudExtension' }
        - { $ref: '#/components/schemas/StorageExtension' }
        - { $ref: '#/components/schemas/ApplicationExtension' }
        - { $ref: '#/components/schemas/SecurityExtension' }
      discriminator:
        propertyName: kind
        mapping:
          IndustrialAsset: '#/components/schemas/IndustrialExtension'
          ComputeAsset: '#/components/schemas/ComputeExtension'
          NetworkAsset: '#/components/schemas/NetworkExtension'
          CloudResource: '#/components/schemas/CloudExtension'
          StorageAsset: '#/components/schemas/StorageExtension'
          ApplicationAsset: '#/components/schemas/ApplicationExtension'
          SecurityAsset: '#/components/schemas/SecurityExtension'
    AssetDetail:
      allOf:
        - { $ref: '#/components/schemas/AssetSummary' }
        - type: object
          required: [capabilities, protocols, interfaces, vendorAttributes, externalReferences, permittedActions]
          properties:
            serialNumber: { $ref: '#/components/schemas/ObservedValue' }
            firmware: { $ref: '#/components/schemas/ObservedValue' }
            hardwareRevision: { $ref: '#/components/schemas/ObservedValue' }
            extension: { $ref: '#/components/schemas/AssetExtension' }
            capabilities:
              type: array
              items: { $ref: '#/components/schemas/Capability' }
            protocols:
              type: array
              items: { $ref: '#/components/schemas/ProtocolBinding' }
            interfaces:
              type: array
              items: { $ref: '#/components/schemas/AssetInterface' }
            vendorAttributes:
              type: array
              items: { $ref: '#/components/schemas/VendorAttribute' }
            externalReferences:
              type: array
              items: { $ref: '#/components/schemas/ExternalReference' }
            businessOwner: { $ref: '#/components/schemas/Owner' }
            maintenanceOwner: { $ref: '#/components/schemas/Owner' }
            permittedActions:
              type: array
              items: { $ref: '#/components/schemas/ActionAvailability' }
    ActionAvailability:
      type: object
      description: Usability hint. Never authoritative.
      required: [action, available]
      properties:
        action: { type: string }
        available: { type: boolean }
        reason: { type: string }
...
    StepUpRequirement:
      type: object
      required: [required]
      properties:
        required: { type: boolean }
        acr: { type: string }
        maxAgeSeconds: { type: integer }
...